Relocate test classes to correct packages.
