+++ src/test/kotlin/fi/hsl/jore4/hastus/util/CsvReaderTest.kt
@@ -1,4 +1,4 @@
-package fi.hsl.jore4.hastus.test
+package fi.hsl.jore4.hastus.util
 
 import fi.hsl.jore4.hastus.data.hastus.ApplicationRecord
 import fi.hsl.jore4.hastus.data.hastus.BlockRecord
@@ -6,7 +6,6 @@ import fi.hsl.jore4.hastus.data.hastus.BookingRecord
 import fi.hsl.jore4.hastus.data.hastus.TripRecord
 import fi.hsl.jore4.hastus.data.hastus.TripStopRecord
 import fi.hsl.jore4.hastus.data.hastus.VehicleScheduleRecord
-import fi.hsl.jore4.hastus.util.CsvReader
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import java.time.LocalDate
